# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic ERP System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface Product {
  id: number;
  name: string;
  sku?: string | null;
  category?: string | null;
  quantity: number;
  cost_price: number;
  sale_price: number;
  low_stock_threshold?: number | null;
  created_at: string;
}

export interface CreateProductInput {
  name: string;
  sku?: string | null;
  category?: string | null;
  quantity: number;
  cost_price: number;
  sale_price: number;
  low_stock_threshold?: number | null;
}

export interface Customer {
  id: number;
  name: string;
  phone?: string | null;
  balance: number;
  created_at: string;
}

export interface CreateCustomerInput {
  name: string;
  phone?: string | null;
  balance?: number;
}

export interface Supplier {
  id: number;
  name: string;
  phone?: string | null;
  balance: number;
  created_at: string;
}

export interface CreateSupplierInput {
  name: string;
  phone?: string | null;
  balance?: number;
}

export interface VoucherInput {
  amount: number;
  description?: string | null;
}

export type SalePaymentType =
  (typeof SalePaymentType)[keyof typeof SalePaymentType];

export const SalePaymentType = {
  cash: "cash",
  credit: "credit",
  partial: "partial",
} as const;

export type SaleStatus = (typeof SaleStatus)[keyof typeof SaleStatus];

export const SaleStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface Sale {
  id: number;
  invoice_no: string;
  customer_name?: string | null;
  customer_id?: number | null;
  payment_type: SalePaymentType;
  total_amount: number;
  paid_amount: number;
  remaining_amount: number;
  status: SaleStatus;
  notes?: string | null;
  created_at: string;
}

export interface SaleItem {
  id: number;
  sale_id: number;
  product_id: number;
  product_name: string;
  quantity: number;
  unit_price: number;
  total_price: number;
}

export type SaleWithItemsPaymentType =
  (typeof SaleWithItemsPaymentType)[keyof typeof SaleWithItemsPaymentType];

export const SaleWithItemsPaymentType = {
  cash: "cash",
  credit: "credit",
  partial: "partial",
} as const;

export type SaleWithItemsStatus =
  (typeof SaleWithItemsStatus)[keyof typeof SaleWithItemsStatus];

export const SaleWithItemsStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface SaleWithItems {
  id: number;
  invoice_no: string;
  customer_name?: string | null;
  customer_id?: number | null;
  payment_type: SaleWithItemsPaymentType;
  total_amount: number;
  paid_amount: number;
  remaining_amount: number;
  status: SaleWithItemsStatus;
  notes?: string | null;
  created_at: string;
  items: SaleItem[];
}

export type CreateSaleInputPaymentType =
  (typeof CreateSaleInputPaymentType)[keyof typeof CreateSaleInputPaymentType];

export const CreateSaleInputPaymentType = {
  cash: "cash",
  credit: "credit",
  partial: "partial",
} as const;

export interface CreateSaleItemInput {
  product_id: number;
  product_name: string;
  quantity: number;
  unit_price: number;
  total_price: number;
}

export interface CreateSaleInput {
  customer_name?: string | null;
  customer_id?: number | null;
  payment_type: CreateSaleInputPaymentType;
  total_amount: number;
  paid_amount: number;
  notes?: string | null;
  items: CreateSaleItemInput[];
}

export type PurchasePaymentType =
  (typeof PurchasePaymentType)[keyof typeof PurchasePaymentType];

export const PurchasePaymentType = {
  cash: "cash",
  credit: "credit",
  partial: "partial",
} as const;

export type PurchaseStatus =
  (typeof PurchaseStatus)[keyof typeof PurchaseStatus];

export const PurchaseStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface Purchase {
  id: number;
  invoice_no: string;
  supplier_name?: string | null;
  supplier_id?: number | null;
  payment_type: PurchasePaymentType;
  total_amount: number;
  paid_amount: number;
  remaining_amount: number;
  status: PurchaseStatus;
  notes?: string | null;
  created_at: string;
}

export interface PurchaseItem {
  id: number;
  purchase_id: number;
  product_id: number;
  product_name: string;
  quantity: number;
  unit_price: number;
  total_price: number;
}

export type PurchaseWithItemsPaymentType =
  (typeof PurchaseWithItemsPaymentType)[keyof typeof PurchaseWithItemsPaymentType];

export const PurchaseWithItemsPaymentType = {
  cash: "cash",
  credit: "credit",
  partial: "partial",
} as const;

export type PurchaseWithItemsStatus =
  (typeof PurchaseWithItemsStatus)[keyof typeof PurchaseWithItemsStatus];

export const PurchaseWithItemsStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface PurchaseWithItems {
  id: number;
  invoice_no: string;
  supplier_name?: string | null;
  supplier_id?: number | null;
  payment_type: PurchaseWithItemsPaymentType;
  total_amount: number;
  paid_amount: number;
  remaining_amount: number;
  status: PurchaseWithItemsStatus;
  notes?: string | null;
  created_at: string;
  items: PurchaseItem[];
}

export type CreatePurchaseInputPaymentType =
  (typeof CreatePurchaseInputPaymentType)[keyof typeof CreatePurchaseInputPaymentType];

export const CreatePurchaseInputPaymentType = {
  cash: "cash",
  credit: "credit",
  partial: "partial",
} as const;

export interface CreatePurchaseItemInput {
  product_id: number;
  product_name: string;
  quantity: number;
  unit_price: number;
  total_price: number;
}

export interface CreatePurchaseInput {
  supplier_name?: string | null;
  supplier_id?: number | null;
  payment_type: CreatePurchaseInputPaymentType;
  total_amount: number;
  paid_amount: number;
  notes?: string | null;
  items: CreatePurchaseItemInput[];
}

export interface Expense {
  id: number;
  category: string;
  amount: number;
  description?: string | null;
  created_at: string;
}

export interface CreateExpenseInput {
  category: string;
  amount: number;
  description?: string | null;
}

export interface Income {
  id: number;
  source: string;
  amount: number;
  description?: string | null;
  created_at: string;
}

export interface CreateIncomeInput {
  source: string;
  amount: number;
  description?: string | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  sale: "sale",
  purchase: "purchase",
  expense: "expense",
  income: "income",
  receipt: "receipt",
  payment: "payment",
} as const;

export interface Transaction {
  id: number;
  type: TransactionType;
  amount: number;
  description?: string | null;
  related_id?: number | null;
  created_at: string;
}

export interface DashboardStats {
  total_sales_today: number;
  total_expenses_today: number;
  total_income_today: number;
  net_profit: number;
  total_customer_debts: number;
  total_supplier_debts: number;
  low_stock_products: Product[];
  recent_transactions: Transaction[];
}