# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Arabic ERP System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: getSuppliers
      tags: [suppliers]
      summary: List all suppliers
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    put:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierInput"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    delete:
      operationId: deleteSupplier
      tags: [suppliers]
      summary: Delete a supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /suppliers/{id}/payment:
    post:
      operationId: createSupplierPayment
      tags: [suppliers]
      summary: Record payment to supplier (payment voucher)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VoucherInput"
      responses:
        "200":
          description: Updated supplier balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /sales:
    get:
      operationId: getSales
      tags: [sales]
      summary: List all sales
      responses:
        "200":
          description: List of sales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"
    post:
      operationId: createSale
      tags: [sales]
      summary: Create a sale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSaleInput"
      responses:
        "201":
          description: Created sale
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Sale"
  /sales/{id}:
    get:
      operationId: getSaleById
      tags: [sales]
      summary: Get sale with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sale details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaleWithItems"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        balance:
          type: number
        created_at:
          type: string
      required: [id, name, balance, created_at]
    CreateSupplierInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        balance:
          type: number
      required: [name]
    VoucherInput:
      type: object
      properties:
        amount:
          type: number
        description:
          type: string
          nullable: true
      required: [amount]
    Sale:
      type: object
      properties:
        id:
          type: integer
        invoice_no:
          type: string
        customer_name:
          type: string
          nullable: true
        customer_id:
          type: integer
          nullable: true
        payment_type:
          type: string
          enum: [cash, credit, partial]
        total_amount:
          type: number
        paid_amount:
          type: number
        remaining_amount:
          type: number
        status:
          type: string
          enum: [paid, partial, unpaid]
        notes:
          type: string
          nullable: true
        created_at:
          type: string
      required: [id, invoice_no, payment_type, total_amount, paid_amount, remaining_amount, status, created_at]
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        sale_id:
          type: integer
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: number
        unit_price:
          type: number
        total_price:
          type: number
      required: [id, sale_id, product_id, product_name, quantity, unit_price, total_price]
    SaleWithItems:
      type: object
      properties:
        id:
          type: integer
        invoice_no:
          type: string
        customer_name:
          type: string
          nullable: true
        customer_id:
          type: integer
          nullable: true
        payment_type:
          type: string
          enum: [cash, credit, partial]
        total_amount:
          type: number
        paid_amount:
          type: number
        remaining_amount:
          type: number
        status:
          type: string
          enum: [paid, partial, unpaid]
        notes:
          type: string
          nullable: true
        created_at:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
      required: [id, invoice_no, payment_type, total_amount, paid_amount, remaining_amount, status, created_at, items]
    CreateSaleInput:
      type: object
      properties:
        customer_name:
          type: string
          nullable: true
        customer_id:
          type: integer
          nullable: true
        payment_type:
          type: string
          enum: [cash, credit, partial]
        total_amount:
          type: number
        paid_amount:
          type: number
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateSaleItemInput"
      required: [payment_type, total_amount, paid_amount, items]
    CreateSaleItemInput:
      type: object
      properties:
        product_id:
          type: integer
        product_name:
          type: string
        quantity:
          type: number
        unit_price:
          type: number
        total_price:
          type: number
      required: [product_id, product_name, quantity, unit_price, total_price]
...