# lib/api-zod/src/generated/types/createSupplierInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic ERP System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSupplierInput {
  name: string;
  phone?: string | null;
  balance?: number;
}