# lib/api-zod/src/generated/types/saleWithItems.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Arabic ERP System API
 * OpenAPI spec version: 0.1.0
 */
import type { SaleItem } from "./saleItem";
import type { SaleWithItemsPaymentType } from "./saleWithItemsPaymentType";
import type { SaleWithItemsStatus } from "./saleWithItemsStatus";

export interface SaleWithItems {
  id: number;
  invoice_no: string;
  customer_name?: string | null;
  customer_id?: number | null;
  payment_type: SaleWithItemsPaymentType;
  total_amount: number;
  paid_amount: number;
  remaining_amount: number;
  status: SaleWithItemsStatus;
  notes?: string | null;
  created_at: string;
  items: SaleItem[];
}